Osa4 part C and D

+++ osa4/blogilista/controllers/blogs.js
@@ -1,17 +1,41 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
-    response.json(blogs.map(blog => blog.toJSON()))
+    const blogData = await Blog.find({})
+      .populate('user', { username: 1, name: 1 })
+
+    response.json(blogData.map(i => i.toJSON()))
 })
 
 blogsRouter.post('/', async (request, response, next) => {
-  const blog = new Blog(request.body)
+  try {
+    const body = request.body
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
-  try{
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'Token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes,
+      user: user._id
+    })
+
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+
+    await user.save()
     response.status(201).json(savedBlog.toJSON())
+
   } catch (exception) {
     next(exception)
   }
@@ -19,6 +43,19 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+    const body = request.body
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'Token missing or invalid' })
+    }
+
+    const blogToDelete = await Blog.findById(request.params.id)
+
+    if (decodedToken.id.toString() !== blogToDelete.user.toString()) {
+      return response.status(401).json({ error: 'Illegal user request: User may only delete own blogs' })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   } catch (exception) {
